cv/pyproject.toml
[project]
name = "cv"
version = "0.1.0"
description = "cv environment"
readme = "README.md"
requires-python = "==3.12"
dependencies = [
    "accelerate>=1.7.0",
    "ensemble-boxes>=1.0.9",
    "fastapi==0.115.12",
    "onnx>=1.12.0",
    "onnxruntime-gpu==1.19.0",
    "onnxslim>=0.1.46",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.1",
    "sahi>=0.11.23",
    "sympy>1.13",
    "tensorrt>=10.10.0.31",
    "torch==2.7.0",
    "torchvision>=0.19.1",
    "ultralytics>=8.3.131",
    "uvicorn[standard]==0.34.2",
]

[tool.uv.workspace]
members = [
    
]


[dependency-groups]
dev = [
    "albumentations==1.4.3",
    "ftfy>=6.3.1",
    "ipython>=9.2.0",
    "jupyter>=1.1.1",
    "ninja>=1.11.1.4",
    "packaging>=25.0",
    "psutil>=7.0.0",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "timm>=1.0.15",
    "tqdm>=4.67.1",
    "transformers>=4.52.3",
]
